requirements.txt
agno==2.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.15.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.5
ddgs==9.10.0
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
duckduckgo_search==8.1.1
edge-tts==7.2.7
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.133.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
feedparser==6.0.12
filelock==3.25.0
fpdf2==2.8.3
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.65.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
libsql-client==0.3.1
libsql-experimental==0.0.55
limits==5.8.0
lxml==5.4.0
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.7.1
newspaper4k==0.9.5
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pgvector==0.4.2
pillow==12.1.1
primp==1.1.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
playwright>=1.49.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rich==14.3.3
pytest==8.3.5
pytest-asyncio==0.25.3
rich-toolkit==0.19.7
rignore==0.7.6
roman-numerals==4.1.0
rsa==4.9.1
sentry-sdk==2.53.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
socksio==1.0.0
soupsieve==2.8.3
Sphinx==9.1.0
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.47
sqlalchemy-libsql==0.2.0
starlette==0.52.1
stripe==14.4.0
tabulate==0.9.0
tavily-python==0.7.22
tenacity==9.1.4
tiktoken==0.12.0
tldextract==5.3.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.6
uvicorn==0.41.0
uvloop==0.22.1
w3lib==2.4.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
yarl==1.22.0
yt-dlp==2025.3.31
argon2-cffi>=23.1.0
